requirements.txt
numpy==1.18.5
opencv-python==4.1.2.30
sockets==1.0.0
pycryptodome==3.12.0
matplotlib